backend/requirements.txt
# Core FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database - Supabase and PostgreSQL
supabase==2.3.0
postgrest==0.12.0
gotrue==2.5.0
storage3==0.6.1
realtime==1.0.4
httpx==0.24.1

# Authentication and Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography>=41.0.0,<42.0.0

# HTTP Requests and Client
requests==2.31.0

# Payment Processing
stripe==7.8.0

# File Handling
aiofiles==23.2.1
Pillow>=10.0.0,<11.0.0

# Environment Variables
python-dotenv==1.0.0

# Data Validation and Serialization
pydantic>=2.5.0,<3.0.0
email-validator==2.1.0

# Date and Time Handling
python-dateutil==2.8.2

# Additional Core Dependencies
typing-extensions>=4.8.0
annotated-types>=0.6.0

# Optional: Development and Testing (uncomment if needed)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# black==23.11.0
# flake8==6.1.0
# isort==5.12.0

# Optional: Enhanced Features (uncomment if needed)
# python-magic==0.4.27        # For file type detection
# redis==5.0.1                # For caching
# celery==5.3.4               # For background tasks
# flower==2.0.1               # For Celery monitoring
